New Ansible role to install ClusterShell
This role installs the [ClusterShell][1] software and defines groups of
nodes based on the groups declared in the Ansible inventory.

[1]: https://cea-hpc.github.io/clustershell/

Configure GitHub Actions to run CI with molecule.

The APT cache is removed during the build of the containers used for the
CI. Since the role needs to install packages, add a preparation step in
molecule to update the APT cache.