feat: Add support of sshpass mode
The sshpass mode is disabled by default.

This change makes sure the sshpass package is installed on the host.
Implement the same logic for the sudo package since the sudo mode is
somewhat similar to sshpass mode.

Since `clustershell_packages` can be a list, make sure the default value
is a list, even if there is a single item. This is required to add items
to the list of packages to install when sshpass mode or sudo mode are
enabled.